add controller page.js and move there all ligic about rendering main page content
